Savepoint before merge
