+++ gym_sts/spaces/observations/utils.py
@@ -1,0 +1,17 @@
+def to_binary_array(n: int, digits: int) -> list[int]:
+    array = [0] * digits
+
+    idx = 0
+    n_copy = n
+    while n_copy > 0:
+        if idx >= digits:
+            raise ValueError(
+                f"{n} is too large to represent with {digits} binary digits"
+            )
+
+        n_copy, r = divmod(n_copy, 2)
+        if r > 0:
+            array[idx] = 1
+        idx += 1
+
+    return array
